requirements.txt
apipkg==1.5
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
execnet==1.7.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.2
pytest-forked==1.3.0
pytest-xdist==2.1.0
pywin32==228
PyYAML==5.3.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0
